Refactor subspaces into component
